Update use checker to use a subclass rather than an instance.
Polymer templates are assumed to be at the same "sytntactic level"
as the class itself, and so should be able to access class internals.
However, the previous code created an instance of the object and
accessed its members through its public interface.

This meant that properties intended to be internal to the element,
but were used in the template, were required to be public and exposed
to a larger audience than intended:

  <template>
    my template includes [[internalOnly]] data
  </template>

  class MyElement ... {
    @property() internalOnly = 'secret';  // this had to be public
  }

This change solves that problem by creating a dummy subclass of the
element for template use checking, and accessing the template's
protected properties.  A user can now write:

  <template>
    my template includes [[internalOnly]] data
  </template>

  class MyElement ... {
    @property() protected internalOnly = 'secret';
  }

This change also adds some semicolons to ensure the generated code
is unambiguous.