Add tests files to .npmignore
Test files shouldn't be included in the npm package.
Also, in the upcoming commits the gts library is added. The gts
reports the error when a file imports a library from devDependencies
and at the same time the file is included in the npm package.
All tests files import chai library and so must be excluded from an
npm package.