Allow to define precise types for elements by id
The polymer allows to define the type of an element in the template:
interface MyElement { // The class MyElement is a polymer element
  $: {
    itemType: HTMLSelectElement; // itemType is the id of an element in the MyElement template
  }
}

Often, an editor (like <select> in the example) are binded to a
non-string property (for example, itemType can be enum). The transpiled
template reports an error (string can't be assigned to enum) in this
case.

This update solves the issue by taking into account exact type for an
element:
interface MyElement { // The class MyElement is a polymer element
  $: {
    itemType: HTMLSelectElement & {value: ItemType};
  }
}

The generated code uses special function:
cons el = __getElementByIdOrTag(container.$, 'itemType', 'select')

It the $.itemType property is defined, the 'el' has the same type as
the property. Otherwise, el has the type HTMLSelectElement (from the
standard HTMLElementTagNameMap).